requirements.txt
boto3
torch
numpy
Pillow
facenet-pytorch
awslambdaric
requests
opencv-python